Remove TitleHeader, replace with MainHeader
