+++ src/pages/Account/Account.js
@@ -1,6 +1,6 @@
 import './Account.scss';
 import { useSelector } from 'react-redux';
-import { Outlet, useResolvedPath } from 'react-router-dom';
+import { Outlet, useNavigate, useResolvedPath } from 'react-router-dom';
 import { useGetAccountQuery } from '../../slices/accountSlice';
 import MainHeader from '../../components/Layout/Headers/MainHeader/MainHeader';
 import DetailsCard from '../../components/DetailsCard/DetailsCard';
@@ -11,13 +11,20 @@ const Account = () => {
     data: account,
     isSuccess
   } = useGetAccountQuery(user?.id);
-
-  const path = useResolvedPath();
-  console.log(path);
+  const { pathname } = useResolvedPath();
+  const navigate = useNavigate();
+  console.log(navigate);
 
   return (
     <>
-      <MainHeader title="Account" />
+      <MainHeader title="Account"
+        left={
+          !pathname === '/account' && (
+            <button className="btn btn--icon btn--transparent btn--back">
+              <i className="fas fa-arrow-left"></i>
+            </button>
+          )
+        } />
       <div className="account-container">
         <DetailsCard
           title={account?.name}
@@ -25,7 +32,6 @@ const Account = () => {
           avatarUrl={account?.avatar_url}
           skeleton={!isSuccess}
         />
-        
         <Outlet context={{ account }}/>
       </div>
     </>
